refactor : 연산자 enum 클래스 이름 변경
