+++ sub_cal/sub_cal/src/main/java/model/Operator.java
@@ -1,71 +1,23 @@
 package model;
 
-import java.util.Stack;
-import java.util.StringTokenizer;
-public class Operator {
-    private Stack<Integer> numberStack = new Stack<>();
-    private Stack<String> operStack = new Stack<>();
+import java.util.Arrays;
+import java.util.Optional;
 
-    public Integer operate(String inputString) {
+public enum Operator {
+    PLUS("+"),
+    MINUS("-"),
+    MULTIPLY("*"),
+    DIVIDE("/");
 
-        StringTokenizer st = new StringTokenizer(inputString);
+    private String operator;
 
-        classify(st);
-
-        return cal();
-    }
-    private void classify(StringTokenizer st){
-
-        while(st.hasMoreTokens()){
-            String word = st.nextToken();
-
-            if(!(OperatorList.getOperator(word).isEmpty())){
-                word = OperatorList.getOperator(word).get().name();
-            }
-
-            switch (word){
-                case "MULTIPLY" : case "DIVIDE":
-                    calPriority(word,Integer.parseInt(st.nextToken()));
-                    break;
-                case "PLUS" : case "MINUS":
-                    operStack.add(word);
-                    break;
-                default:
-                    if(operStack.isEmpty() || operStack.pop().equals("PLUS")){
-                        numberStack.push(Integer.parseInt(word));
-                        break;
-                    }
-
-                    numberStack.push(-1*Integer.parseInt(word));
-                    break;
-            }
-        }
-
-
+    public static Optional<Operator> getOperator(String inputOperator) {
+        return Arrays.stream(values())
+                .filter(o -> o.operator.equals(inputOperator))
+                .findFirst();
     }
 
-    private Integer cal(){
-        while(true){
-            if(numberStack.size() == 1) break;
-
-            int num2 = numberStack.pop();
-            int num1 = numberStack.pop();
-
-            numberStack.push(num1 + num2);
-
-
-        }
-
-        return numberStack.pop();
-    }
-
-    private void calPriority(String word,int num2) {
-        int num1 = numberStack.pop();
-
-        if(word.equals("MULTIPLY")) numberStack.push(num1*num2);
-
-        else numberStack.push(num1/num2);
-
-
+     Operator(String operator) {
+        this.operator = operator;
     }
 }
